requirements.txt
asgiref==3.11.0
certifi==2025.11.12
distlib==0.4.0
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
packaging==25.0
pipenv==2025.0.4
platformdirs==4.5.0
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.2
virtualenv==20.35.4
